requirements.txt
Flask==1.0.2
line-bot-sdk==1.8.0
dropbox==9.1.0
datetime==4.3
numpy==1.15.1